Moved DialogBlocker to be a part of Dialog
