+++ src/App.js
@@ -8,7 +8,8 @@ import ScrmblButton from './components/ScrmblButton';
 import KeyboardListener from './components/KeyboardListener';
 import getDailyScrmbl from './getDailyScrmbl';
 import GameOver from './components/GameOver';
-import { motion, AnimatePresence } from 'framer-motion';
+
+import { AnimatePresence } from 'framer-motion';
 
 const Content = styled.div`
   display: flex;
@@ -54,19 +55,6 @@ const ScrmblContainer = styled.div`
   margin-top: auto;
   background: #333;
   border-radius: 100px;
-`;
-
-const DialogBlocker = styled(motion.div)`
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: #000000dd;
-  z-index: 99;
 `;
 
 function App() {
@@ -220,22 +208,19 @@ function App() {
   return (
     <Content>
       <AnimatePresence>
-        {(helpActive || gameOver) && <DialogBlocker
-          initial={{ opacity: 0}}
-          animate={{ opacity: 1}}
-          exit={{ opacity: 0}}
-        >
-          {helpActive && <Welcome
+        {helpActive ?
+          <Welcome
             close={() => setHelpActive(false)}
             startClock={startClock}
             gameOver={gameOver}
-          />}
-          {gameOver && <GameOver
+          />
+        : gameOver &&
+          <GameOver
             scrmblsLeft={scrmblsLeft}
             elapsedSeconds={elapsedSeconds}
             word={word}
-          />}
-        </DialogBlocker>}
+          />
+        }
       </AnimatePresence>
       <Header>
         <Branding>
